group cli tx files in tx.go and cli query files in query.go
