--- x/slashrefund/client/cli/query_deposit_pool.go
@@ -1,73 +1,0 @@
-package cli
-
-import (
-	"context"
-
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/made-in-block/slash-refund/x/slashrefund/types"
-	"github.com/spf13/cobra"
-)
-
-func CmdListDepositPool() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "list-deposit-pool",
-		Short: "list all deposit_pool",
-		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx := client.GetClientContextFromCmd(cmd)
-
-			pageReq, err := client.ReadPageRequest(cmd.Flags())
-			if err != nil {
-				return err
-			}
-
-			queryClient := types.NewQueryClient(clientCtx)
-
-			params := &types.QueryAllDepositPoolRequest{
-				Pagination: pageReq,
-			}
-
-			res, err := queryClient.DepositPoolAll(context.Background(), params)
-			if err != nil {
-				return err
-			}
-
-			return clientCtx.PrintProto(res)
-		},
-	}
-
-	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
-	flags.AddQueryFlagsToCmd(cmd)
-
-	return cmd
-}
-
-func CmdShowDepositPool() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "show-deposit-pool [operator-address]",
-		Short: "shows a deposit_pool",
-		Args:  cobra.ExactArgs(1),
-		RunE: func(cmd *cobra.Command, args []string) (err error) {
-			clientCtx := client.GetClientContextFromCmd(cmd)
-
-			queryClient := types.NewQueryClient(clientCtx)
-
-			argOperatorAddress := args[0]
-
-			params := &types.QueryGetDepositPoolRequest{
-				OperatorAddress: argOperatorAddress,
-			}
-
-			res, err := queryClient.DepositPool(context.Background(), params)
-			if err != nil {
-				return err
-			}
-
-			return clientCtx.PrintProto(res)
-		},
-	}
-
-	flags.AddQueryFlagsToCmd(cmd)
-
-	return cmd
-}
