+++ x/slashrefund/client/cli/tx.go
@@ -7,7 +7,10 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/cosmos/cosmos-sdk/client"
-	// "github.com/cosmos/cosmos-sdk/client/flags"
+	"github.com/cosmos/cosmos-sdk/client/flags"
+	"github.com/cosmos/cosmos-sdk/client/tx"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
 	"github.com/made-in-block/slash-refund/x/slashrefund/types"
 )
 
@@ -34,6 +37,102 @@ func GetTxCmd() *cobra.Command {
 	cmd.AddCommand(CmdWithdraw())
 	cmd.AddCommand(CmdClaim())
 	// this line is used by starport scaffolding # 1
+
+	return cmd
+}
+
+func CmdDeposit() *cobra.Command {
+
+	cmd := &cobra.Command{
+		Use:   "deposit [validator-address] [amount]",
+		Short: "Broadcast message deposit",
+		Args:  cobra.ExactArgs(2),
+		RunE: func(cmd *cobra.Command, args []string) (err error) {
+			argValidatorAddress := args[0]
+			argAmount := args[1]
+
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			amount, err := sdk.ParseCoinNormalized(argAmount)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgDeposit(
+				clientCtx.GetFromAddress().String(),
+				argValidatorAddress,
+				amount,
+			)
+
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
+func CmdWithdraw() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "withdraw [validator-address] [amount]",
+		Short: "Broadcast message withdraw",
+		Args:  cobra.ExactArgs(2),
+		RunE: func(cmd *cobra.Command, args []string) (err error) {
+			argValidatorAddress := args[0]
+			argAmount := args[1]
+
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			amount, err := sdk.ParseCoinNormalized(argAmount)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgWithdraw(
+				clientCtx.GetFromAddress().String(),
+				argValidatorAddress,
+				amount,
+			)
+
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
+
+	return cmd
+}
+
+func CmdClaim() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "claim [validator-address]",
+		Short: "Claim refund from the slashed validator",
+		Args:  cobra.ExactArgs(1),
+		RunE: func(cmd *cobra.Command, args []string) (err error) {
+			argValidatorAddress := args[0]
+
+			clientCtx, err := client.GetClientTxContext(cmd)
+			if err != nil {
+				return err
+			}
+
+			msg := types.NewMsgClaim(
+				clientCtx.GetFromAddress().String(),
+				argValidatorAddress,
+			)
+
+			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
+		},
+	}
+
+	flags.AddTxFlagsToCmd(cmd)
 
 	return cmd
 }
