--- x/slashrefund/client/cli/tx_claim.go
@@ -1,36 +1,0 @@
-package cli
-
-import (
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/cosmos/cosmos-sdk/client/tx"
-	"github.com/made-in-block/slash-refund/x/slashrefund/types"
-	"github.com/spf13/cobra"
-)
-
-func CmdClaim() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "claim [validator-address]",
-		Short: "Claim refund from the slashed validator",
-		Args:  cobra.ExactArgs(1),
-		RunE: func(cmd *cobra.Command, args []string) (err error) {
-			argValidatorAddress := args[0]
-
-			clientCtx, err := client.GetClientTxContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			msg := types.NewMsgClaim(
-				clientCtx.GetFromAddress().String(),
-				argValidatorAddress,
-			)
-
-			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
-		},
-	}
-
-	flags.AddTxFlagsToCmd(cmd)
-
-	return cmd
-}
