--- x/slashrefund/client/cli/tx_deposit.go
@@ -1,45 +1,0 @@
-package cli
-
-import (
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/cosmos/cosmos-sdk/client/tx"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/made-in-block/slash-refund/x/slashrefund/types"
-	"github.com/spf13/cobra"
-)
-
-func CmdDeposit() *cobra.Command {
-
-	cmd := &cobra.Command{
-		Use:   "deposit [validator-address] [amount]",
-		Short: "Broadcast message deposit",
-		Args:  cobra.ExactArgs(2),
-		RunE: func(cmd *cobra.Command, args []string) (err error) {
-			argValidatorAddress := args[0]
-			argAmount := args[1]
-
-			clientCtx, err := client.GetClientTxContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			amount, err := sdk.ParseCoinNormalized(argAmount)
-			if err != nil {
-				return err
-			}
-
-			msg := types.NewMsgDeposit(
-				clientCtx.GetFromAddress().String(),
-				argValidatorAddress,
-				amount,
-			)
-
-			return tx.GenerateOrBroadcastTxCLI(clientCtx, cmd.Flags(), msg)
-		},
-	}
-
-	flags.AddTxFlagsToCmd(cmd)
-
-	return cmd
-}
